pyproject.toml
[tool.poetry]
name = "ai-book"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
uvicorn = "^0.30.6"
fastapi = "^0.112.2"
langchain-core = "^0.2.37"
langchain-openai = "^0.1.23"
langchain-community = "^0.2.15"
redis = "^5.0.8"
google-cloud-texttospeech = "^2.17.2"
python-multipart = "^0.0.9"
jq = "^1.8.0"
rank-bm25 = "^0.2.2"
aioredis = "^2.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
